use 32-bit numbers instead of 64-bit numbers
we're storing a 16-bit unsigned values, so we need an i32 to get all the
way up there. Annoying that we need the extra space, but that's just the
way the cookie crumbles. It wouldn't make any sense to model these
fields using signed integers.

+++ beeps-server/migrations/20250119041226_minutes_per_pings.sql
@@ -5,8 +5,8 @@ CREATE TABLE IF NOT EXISTS minutes_per_pings (
     minutes_per_ping BIGINT NOT NULL,
     -- clock
     clock TIMESTAMPTZ NOT NULL,
-    counter BIGINT NOT NULL,
-    node_id BIGINT NOT NULL,
+    counter INTEGER NOT NULL,
+    node_id INTEGER NOT NULL,
     -- references
     FOREIGN KEY (document_id) REFERENCES accounts (id) ON DELETE CASCADE,
     UNIQUE (document_id, clock, counter, node_id)
